Initial process to build program
Created a loop to keep the program running until the user enters x/X
char.

I'm using the `commands` method to match the user command and using it
to help users to understand what are the parameters accepted and how
they work, as for me that's a requirement and would be my extra
functionality to this program.

TODO: I also must check that "I M N" had run befor other commands,
otherwise doesn't make sense.