Adds Image spec
- One of the things I had in my head was the min and max of the image:
`1 <= M , N <= 250`

- Then I found it would be interesting to see if it will generate all the
pixels for the image.

- And for last I needed to know if it would show the result.

I tried to come with a better name other then 'image.pixels ...
have(..).pixels', seems it might be true that it's hard to come
with good names.

My other though for the show was using `<<-image` but the identation was
messing with the result, so my last approach was a one-line string.