Implement DrawHorizontal Command
Follow the SRP by having one command responsible for one action,
in this case this command is responsible to add characters on the X axis
of the Bitmap.

The coordinates in this program start from the top left, meaning the
zero coordianate starts on the top left and not on the bottom left.

E.g:

     0 1 2 3  X
   +---------->
 0 | O O O O
 1 | O O O O
 2 | O O O O
 3 | O O O O
 Y v