Extract validation of coordinates to own validator
This commit extracts the logic to validate the boundaries of the bitmap
into it's own validator.

FillRegion class now accepts a validator, meaning that if in the future
there's a need to validate something else we can plug in another
validator as long as it follows the same implementation of
CoordinatesValidator, which is also the default validator at the moment.

This commit also introduces a Validators::Base, which holds common
behaviour between different validators.